Fire output as a grid
The fire module has now a grid based output (year of last burn for each
pixel).
The state change output uses now the cell index instead of the
environment Id as the Id.

+++ SVDModel/SVDCore/modules/fire/firemodule.h
@@ -67,8 +67,8 @@ private:
     double mExtinguishProb;
 
     // index of variables
-    int miBurnProbability;
-    int miHighSeverity;
+    size_t miBurnProbability;
+    size_t miHighSeverity;
 
     // fire statistics
     std::vector< SFireStat > mStats;
